# Use synchronized instead of reentrant lock in explicit bucket histogram