@jack-berg 2024-03-21
Updated to use a dedicated lock, added this advice to the best practices guide